[refactor] domain => infraへの依存をディレクトリレベルで取り除いた
